Move backend and download services into subpackages.
