+++ Sebastiaanschool/src/main/java/nl/sebastiaanschool/contact/app/data/GrabBag.java
@@ -8,14 +8,14 @@ import java.util.Locale;
 
 import nl.sebastiaanschool.contact.app.BuildConfig;
 
-class GrabBag {
+public class GrabBag {
     /**
      * Creates a user-agent string that should generally be RFC compliant. If the device constants
      * contain context-invalid characters these are used as-is and RFC compliance will suffer.
      * @param context used to obtain some device data.
      * @return e.g. "Sebastiaanschool/2.0.0 (LGE; Nexus 5; Android 6.1 SDK 23; nl_NL)"
      */
-    static String constructUserAgent(@NonNull Context context) {
+    public static String constructUserAgent(@NonNull Context context) {
         return  new StringBuilder(256)
                 .append("Sebastiaanschool/").append(BuildConfig.VERSION_NAME)
                 .append(" (").append(Build.MANUFACTURER).append("; ").append(Build.MODEL)
